通知メッセージを整備する Fixes #26
